feat: 切り出したうえでupdate関数を追加
